feat(debug): render adaptive learner as animated node graphs
Make the inspector a living score-merge / constellation mesh and keep
resolved outcomes visible in the learning pulse under Home noise.

+++ core/data/src/main/java/cx/aswin/boxcast/core/data/ranking/RankingFeedbackRepository.kt
@@ -179,12 +179,9 @@ class RankingFeedbackRepository private constructor(
 
         fun getIfInitialized(): RankingFeedbackRepository? = instance
 
-        private fun progressRatio(
-            listenSeconds: Long,
-            durationSeconds: Long,
-        ): Double {
-            if (durationSeconds <= 0) return 0.0
-            return listenSeconds.toDouble() / durationSeconds
+        private fun progressRatio(listenSeconds: Long, durationSeconds: Long): Double {
+            if (durationSeconds <= 0L) return 0.0
+            return listenSeconds.toDouble() / durationSeconds.toDouble()
         }
     }
 }
